Restructures tests.
