+++ tests/conftest.py
@@ -2,7 +2,6 @@
 
 '''
 
-
 import numpy as np
 import pytest
 import torch
@@ -23,7 +22,9 @@ def arguments():
 
 @pytest.fixture
 def data_class():
+
     class DummyData():
+
         def __init__(self, dim):
             d_ = np.arange(50 * dim)
             d_ = d_.reshape((50, dim)).astype('float32') * 0.01
@@ -37,14 +38,14 @@ def data_class():
                 self.i = 0
                 raise StopIteration
 
-            d = self._data[self.i * self.bs: (self.i + 1) * self.bs]
+            d = self._data[self.i * self.bs:(self.i + 1) * self.bs]
             self.i += 1
             return d
 
         def __getitem__(self, item):
             if item != 'test':
                 raise KeyError(item)
-            return self._data[self.i * self.bs: (self.i + 1) * self.bs]
+            return self._data[self.i * self.bs:(self.i + 1) * self.bs]
 
         def reset(self, *args, **kwargs):
             self.i = 0
@@ -64,8 +65,10 @@ def model_class(arguments):
     '''.format(**arguments)
 
     class TestModel(ModelPlugin):
+
         def build(self, a=17, b=19):
             self.nets.net = FullyConnectedNet(a, b)
+
         build.__doc__ = _build_doc
 
         def routine(self, A):
@@ -74,12 +77,15 @@ def model_class(arguments):
 
             self.losses.net = output.sum()
             self.results.output = output.sum().item()
+
     return TestModel
 
 
 @pytest.fixture
 def model_class_with_submodel(model_class):
+
     class TestModel2(ModelPlugin):
+
         def __init__(self, sub_contract=None, *args, **kwargs):
             super().__init__(*args, **kwargs)
             self.submodel = model_class(contract=sub_contract)
@@ -119,7 +125,9 @@ def model_class_with_submodel(model_class):
 
 @pytest.fixture
 def model_class_with_submodel_2(model_class):
+
     class TestModel3(ModelPlugin):
+
         def __init__(self, sub_contract1, sub_contract2, *args, **kwargs):
             super().__init__(*args, **kwargs)
             self.submodel1 = model_class(contract=sub_contract1)
@@ -169,16 +177,90 @@ def model_with_submodel(model_class_with_submodel, data_class):
 
     contract = dict(inputs=dict(B='test'))
     sub_contract = dict(
-        kwargs=dict(a='d'),
-        nets=dict(net='net2'),
-        inputs=dict(A='test')
-    )
+        kwargs=dict(a='d'), nets=dict(net='net2'), inputs=dict(A='test'))
 
-    model = model_class_with_submodel(sub_contract=sub_contract,
-                                      contract=contract)
+    model = model_class_with_submodel(
+        sub_contract=sub_contract, contract=contract)
     model._data = data
     model.submodel._data = data
 
     model.kwargs.update(**kwargs)
 
     return model
+
+
+@pytest.fixture
+def args():
+    from argparse import Namespace
+    import torch.nn as nn
+    return Namespace(
+        classifier_args={'dropout': 0.2},
+        classifier_type='convnet',
+        clean=False,
+        command=None,
+        config_file=None,
+        criterion=nn.CrossEntropyLoss(),
+        device=0,
+        load_models=None,
+        meta=None,
+        name=None,
+        out_path=None,
+        reload=None,
+        reloads=None,
+        verbosity=1,
+        load_networks=False,
+        **{
+            'data.batch_size': 128,
+            'data.copy_to_local': True,
+            'data.data_args': None,
+            'data.inputs': {
+                'inputs': 'images'
+            },
+            'data.n_workers': 4,
+            'data.shuffle': True,
+            'data.skip_last_batch': False,
+            'data.source': 'CIFAR10',
+            'optimizer.clipping': None,
+            'optimizer.learning_rate': 0.001,
+            'optimizer.model_optimizer_options': None,
+            'optimizer.optimizer': 'Adam',
+            'optimizer.optimizer_options': None,
+            'optimizer.weight_decay': None,
+            'train.archive_every': 10,
+            'train.epochs': 1,
+            'train.eval_during_train': True,
+            'train.eval_only': False,
+            'train.quit_on_bad_values': True,
+            'train.save_on_best': 'losses.classifier',
+            'train.save_on_highest': None,
+            'train.save_on_lowest': None,
+            'train.test_mode': 'test',
+            'train.train_mode': 'train'
+        })
+
+
+@pytest.fixture
+def classifier_modified():
+    from cortex.built_ins.models.utils import update_encoder_args
+    from cortex.built_ins.models.classifier import ImageClassification
+
+    class ClassifierModified(ImageClassification):
+        defaults = dict(
+            data=dict(batch_size=128, inputs=dict(inputs='images')),
+            optimizer=dict(optimizer='Adam', learning_rate=1e-3),
+            train=dict(epochs=200, save_on_best='losses.classifier'),
+            model=dict(classifier_type='resnet'))
+
+        def build(self,
+                  classifier_type='convnet',
+                  classifier_args=dict(dropout=0.2)):
+            classifier_args = classifier_args or {}
+            shape = self.get_dims('x', 'y', 'c')
+            dim_l = self.get_dims('labels')
+            Encoder, args = update_encoder_args(
+                shape, model_type=classifier_type, encoder_args=classifier_args)
+            args.update(**classifier_args)
+            classifier = Encoder(shape, dim_out=dim_l, **args)
+            self.nets.classifier = classifier
+
+    return ClassifierModified()
